requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
idna==2.10
    # via requests
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements.in
python-dateutil==2.8.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
toml==0.10.2
    # via pep517
urllib3==1.26.5
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
